Better permission support
Allow permissions to be controled in a number of different ways.
filterMusic, filterPhotos, and JSON. The filter options have not changed
and can be modified via plex.tv for any user. JSON is new and requires
the use of a JSON formatted file. PlexAuth/inc/permissions.json for an
example. Use http://www.jsoneditoronline.org/ to ensure json format is
correct. Thanks @Burry for the start code and idea.